feat: reorganize project structure for v0.1 milestone (#207)
Reorganized project structure with professional layout:
- Moved documentation to docs/ directory
- Organized scripts in scripts/ directory  
- Created config/ for configuration files
- Implemented backward compatibility via compat/ shims
- Preserved git history using git mv for all file movements

All references updated and functionality maintained.
