+++ README.md
@@ -179,7 +179,21 @@ gadugi/
 │   ├── Memory.md                   # AI assistant persistent memory
 │   └── workflows/                  # GitHub Actions workflows
 ├── prompts/                        # Prompt templates
-├── manifest.yaml                   # Agent registry and versions
+├── docs/                           # Documentation
+│   ├── architecture/
+│   │   ├── AGENT_HIERARCHY.md      # Agent system hierarchy
+│   │   └── SYSTEM_DESIGN.md        # System design documentation
+│   └── templates/
+│       └── CLAUDE_TEMPLATE.md      # Claude instruction template
+├── scripts/                        # Utility scripts
+│   ├── claude                      # Claude CLI executable
+│   ├── claude-worktree-manager.sh  # Worktree management
+│   └── launch-claude-*.sh          # Launch helpers
+├── config/                         # Configuration files
+│   ├── manifest.yaml               # Agent registry and versions
+│   └── vscode-claude-terminals.json # VSCode configuration
+├── compat/                         # Compatibility shims for legacy imports
+├── types/                          # Type definitions and stubs
 ├── CLAUDE.md                       # Project-specific AI instructions
 ├── claude-generic-instructions.md  # Generic Claude Code best practices
 ├── LICENSE                         # MIT License
