+++ compat/__init__.py
@@ -1,0 +1,7 @@
+"""
+Compatibility shims for legacy imports.
+
+This package contains compatibility shims that redirect imports to their
+canonical implementations in .claude/shared/. This allows legacy code to
+continue working while maintaining a single source of truth.
+"""
