+++ compat/github_operations.py
@@ -17,7 +17,10 @@ from pathlib import Path
 from types import ModuleType
 
 _IMPL_PATH = (
-    Path(__file__).resolve().parent / ".claude" / "shared" / "github_operations.py"
+    Path(__file__).resolve().parent.parent
+    / ".claude"
+    / "shared"
+    / "github_operations.py"
 )
 
 if not _IMPL_PATH.is_file():
