+++ config/manifest.yaml
