+++ config/vscode-claude-terminals.json
