+++ docs/architecture/AGENT_HIERARCHY.md
