+++ scripts/claude
@@ -1,0 +1,1 @@
+.claude
