+++ scripts/claude-worktree-manager.sh
