+++ scripts/launch-claude-vscode.py
