+++ scripts/restart-claude-worktrees.sh
