+++ types/components/prompt_generator.pyi
