+++ types/components/worktree_manager.pyi
