+++ types/docker-stubs/__init__.pyi
