+++ types/memory_compactor.pyi
